requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
PyPDF2
requests
google-generativeai
